feat: add errors message
