+++ src/graph/fragments/debug.ts
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-import { QueryBrowserErrorLogs, QueryServiceLogs, QueryLogsByKeywords } from '../fragments/log';
-
-export const queryBrowserErrorLogs = `query queryBrowserErrorLogs(${QueryBrowserErrorLogs.variable}) {
-  ${QueryBrowserErrorLogs.query}}`;
-export const queryServiceLogs = `query queryLogs(${QueryServiceLogs.variable}) {${QueryServiceLogs.query}}`;
-export const queryLogsByKeywords = `query queryLogsByKeywords {${QueryLogsByKeywords.query}}`;
+export const QueryLogTest = {
+  variable: '$requests: LogTestRequest!',
+  query: `test(requests: $requests) { log { content, serviceName, serviceInstanceName, endpointName,
+    traceId, timestamp, contentType, content, tags { key, value }}
+  metrics { name, value, timestamp, tags { key, value }} }`,
+};
