+++ src/graph/fragments/log.ts
@@ -63,10 +63,3 @@ export const QueryLogsByKeywords = {
   query: `
   support: supportQueryLogsByKeywords`,
 };
-
-export const QueryLogTest = {
-  variable: '$requests: LogTestRequest!',
-  query: `test(requests: $requests) { log { content, serviceName, serviceInstanceName, endpointName,
-    traceId, timestamp, contentType, content, tags { key, value }}
-  metrics { name, value, timestamp, tags { key, value }} }`,
-};
