+++ src/graph/query/debug.ts
@@ -15,9 +15,6 @@
  * limitations under the License.
  */
 
-import { QueryBrowserErrorLogs, QueryServiceLogs, QueryLogsByKeywords } from '../fragments/log';
+import { QueryLogTest } from '../fragments/debug';
 
-export const queryBrowserErrorLogs = `query queryBrowserErrorLogs(${QueryBrowserErrorLogs.variable}) {
-  ${QueryBrowserErrorLogs.query}}`;
-export const queryServiceLogs = `query queryLogs(${QueryServiceLogs.variable}) {${QueryServiceLogs.query}}`;
-export const queryLogsByKeywords = `query queryLogsByKeywords {${QueryLogsByKeywords.query}}`;
+export const queryLogTest = `query test(${QueryLogTest.variable}) {${QueryLogTest.query}}`;
